Version 7.1.17 (27.07.2022) - UPDATED: Calculations of round timing

+++ net.clanwolf.starmap.server/src/main/java/net/clanwolf/starmap/server/persistence/pojos/SeasonPOJO.java
@@ -68,15 +68,15 @@ public class SeasonPOJO extends Pojo {
 	private Date startDate;
 
 	@Column(name = "DaysInRound")
-	private Long daysInRound;
+	private Double daysInRound;
 
 	@SuppressWarnings("unused")
-	public Long getDaysInRound() {
+	public Double getDaysInRound() {
 		return daysInRound;
 	}
 
 	@SuppressWarnings("unused")
-	public void setDaysInRound(Long daysInRound) {
+	public void setDaysInRound(Double daysInRound) {
 		this.daysInRound = daysInRound;
 	}
